chore: separate json-schema-test-suite related items from validation
- makes it cleaner and easier to understand whats going on
