+++ next/test/json-schema-test-suite/constants.ts
