+++ next/test/json-schema-test-suite/json-schema-test-suite-tracker.js
@@ -2,14 +2,17 @@ import fs from 'node:fs'
 import path from 'node:path'
 import { fileURLToPath } from 'node:url'
 
-import { JSON_SCHEMA_SUITE_FAILED_TESTS_FILE_NAME } from './constants.js'
+import { JSON_SCHEMA_SUITE_FAILED_TESTS_FILE_NAME } from './constants'
 
 const __filename = fileURLToPath(import.meta.url)
 const __dirname = path.dirname(__filename)
 
 // Save newly failed tests
 function saveFailedTests(failedTests) {
-  fs.writeFileSync(path.join(__dirname, JSON_SCHEMA_SUITE_FAILED_TESTS_FILE_NAME), JSON.stringify({ failedTests }, null, 2))
+  fs.writeFileSync(
+    path.join(__dirname, JSON_SCHEMA_SUITE_FAILED_TESTS_FILE_NAME),
+    `${JSON.stringify({ failedTests }, null, 2)}\n`,
+  )
 }
 
 /**
